Create an undoManager if none exists
10.12 stopped creating undo managers for MOCs automatically.

Before the @jerrykrinock commits, the karelia code stopped the document
from receiving a non-MOC undoManager by overriding setUndoManager to
be a no-op. Now, the setter acts as expected, and the document receives
its default undoManager from AppKit, but then this manager is soon
replaced with the MOC's undoManager during setManagedObjectContext:.

Comments from @theMikeSwan indicate that calls to .undoManager cause a
crash in 10.14 (maybe need to be wrapped in performBlockAndWait?) but I
have not seen these crashes anywhere.