Remove persistent store when closing document
If the persistent store is not manually removed, the file will be
unlinked while the SQLite file descriptor is still open, resulting
in these errors:

    BUG IN CLIENT OF libsqlite3.dylib: database integrity compromised by
    API violation: vnode unlinked while in use

Note that this change may result in crashes if client code attempts to
modify managed objects after document close (e.g. in dealloc methods
that access @sythesized relationships to other managed objects). This
client code will need to be fixed.