Make the NSPersistentStoreCoordinator an ivar
The deadlock described in jerrykrinock#8 can be traced to a main-queue
context attempting to `performBlockAndWait:` while the application's
main thread is blocked by the Version Browser. This patch creates a
document-owned coordinator (assigned to contexts later) in order to move
some logic from the context's `performBlockAndWait:` to the
coordinator's `performBlockAndWait:`, which does not appear to deadlock.

Having the document own the coordinator lets us remove some bits of
indirection and generally simplify the code. My only concern with this
approach is that [NSPersistentStoreCoordinator peformBlockAndWait:]
is not available pre-10.10, and so adding the persistent store on
a secondary thread may cause issues with context notifications that
expect it to happen on the main thread, as described in the comment
for `configurePersistentStoreCoordinatorForURL:`.