Pre-10.10 fixes when adding/removing persistent store
`[NSPersistentStoreCoordinator performBlockAndWait:]` is only available
on 10.10 and later. On earlier platforms, fall back to the existing
context's `performBlockAndWait:`. However, be careful not to create
the context as a side-effect (which can deadlock while waiting for
access to the main thread); therefore use `_managedObjectContext`
instead of `self.managedObjectContext`. If no context exists, then
adding/removing the persistent store on any thread should be fine.